Finished apartment scraper (for now lmao)
Finished apartment scraper. Three python files: GetUrls.py: gets the urls of the apartment pages on apartments.com, GetAmenities.py: gets all the necesarry information from the apartment pages (rent, sqft, numbeds, numbaths, ameneties), an getinfo.py, which builds a dictionary of the apartment names as the keys, and the apartment rooms and the information for those rooms as values. I also included a csv file so the data is easy to see.